[FIX] Dedicated feed-all button instead of feeder machine hijack
Player feedback: overloading the feeder machine click with the golden
asset feed-all action removed control over opening the crafting modal.

- Revert FeederMachine to always open the crafting modal (lightning
  bolt and feed-all click handling removed)
- New FeedAllButton beside the feeder machine: grinder icon with a
  lightning bolt, pulsing when actionable, dimmed when nothing is
  eligible; hidden without a covering golden asset
- Same eligibility/double-tap guards and animals.fedAll event as before
