+++ src/features/feederMachine/FeederMachine.tsx
@@ -1,108 +1,21 @@
 import { SUNNYSIDE } from "assets/sunnyside";
 import { PIXEL_SCALE } from "features/game/lib/constants";
-import { useAppTranslation } from "lib/i18n/useAppTranslations";
-import React, { useContext, useEffect, useMemo, useState } from "react";
-import { useSelector } from "@xstate/react";
-import { Context } from "features/game/GameProvider";
-import type { MachineState } from "features/game/lib/gameMachine";
-import { makeAnimalBuildingKey } from "features/game/lib/animals";
-import type { AnimalBuildingType } from "features/game/types/animals";
-import {
-  getCoveredAnimalTypes,
-  getFeedAllTargets,
-} from "features/game/events/landExpansion/feedAllAnimals";
-import { useSound } from "lib/utils/hooks/useSound";
-import { getValues } from "lib/object";
-import classNames from "classnames";
+import React, { useState } from "react";
 import { FeederMachineModal } from "./FeederMachineModal";
 
-const _state = (state: MachineState) => state.context.state;
-
 interface Props {
-  // A golden asset covering this building turns the machine into a
-  // one-click feed-all trigger (shown with a lightning bolt). The crafting
-  // modal stays reachable whenever medicine or feed is still needed.
-  building: AnimalBuildingType;
+  building: "Hen House" | "Barn";
 }
 
 export const FeederMachine: React.FC<Props> = ({ building }) => {
-  const { t } = useAppTranslation();
   const [showFeederMachineModal, setFeederMachineModal] = useState(false);
   const feederMachineImage = SUNNYSIDE.building.feederMachine;
-
-  const { gameService } = useContext(Context);
-  const game = useSelector(gameService, _state);
-  const { play: playFeedAnimal } = useSound("feed_animal");
-
-  // Bumped when the soonest sleeping animal wakes so eligibility re-evaluates
-  const [, setWakeTick] = useState(0);
-
-  const covered = useMemo(
-    () => getCoveredAnimalTypes({ state: game, building }),
-    [building, game],
-  );
-
-  const targets = getFeedAllTargets({ state: game, building });
-  const eligibleCount =
-    covered.length > 0
-      ? targets.toClaim.length + targets.toCure.length + targets.toFeed.length
-      : 0;
-
-  const allWakeTimes = useMemo(
-    () =>
-      getValues(game[makeAnimalBuildingKey(building)].animals)
-        .filter((animal) => covered.includes(animal.type))
-        .map((animal) => animal.awakeAt),
-    [building, game, covered],
-  );
-  const nextWakeAt = Math.min(
-    ...allWakeTimes.filter(
-      // eslint-disable-next-line react-hooks/purity
-      (awakeAt) => awakeAt > Date.now(),
-    ),
-  );
-
-  useEffect(() => {
-    if (!isFinite(nextWakeAt)) return;
-
-    const timeout = setTimeout(
-      () => setWakeTick((tick) => tick + 1),
-      nextWakeAt - Date.now() + 100,
-    );
-
-    return () => clearTimeout(timeout);
-  }, [nextWakeAt]);
-
-  const hasGoldenAsset = covered.length > 0;
-  // Sick animals the action cannot cure are simply skipped (manual care);
-  // they do not block feeding the rest of the building.
-  const canFeedAll = hasGoldenAsset && eligibleCount > 0;
-
-  const handleClick = () => {
-    if (covered.length > 0) {
-      // Re-derive eligibility from the live machine snapshot rather than
-      // the render-scoped state, which can go stale (e.g. a double-tap
-      // before React re-renders after the first click consumed every
-      // target).
-      const { toClaim, toCure, toFeed } = getFeedAllTargets({
-        state: gameService.getSnapshot().context.state,
-        building,
-      });
-      if (toClaim.length + toCure.length + toFeed.length > 0) {
-        gameService.send({ type: "animals.fedAll", building });
-        playFeedAnimal();
-        return;
-      }
-    }
-
-    setFeederMachineModal(true);
-  };
 
   return (
     <>
       <div
         className="relative cursor-pointer hover:img-highlight"
-        onClick={handleClick}
+        onClick={() => setFeederMachineModal(true)}
       >
         <img
           src={SUNNYSIDE.animalFoods.grinder}
@@ -112,19 +25,6 @@ export const FeederMachine: React.FC<Props> = ({ building }) => {
           // Hover parent
           className="absolute top-0 -right-4 z-20"
         />
-        {hasGoldenAsset && (
-          <img
-            src={SUNNYSIDE.icons.lightning}
-            alt={t("animals.feedAll")}
-            className={classNames(
-              "absolute z-30 top-0 -right-4 pointer-events-none img-highlight",
-              canFeedAll ? "animate-pulsate" : "grayscale opacity-50",
-            )}
-            style={{
-              width: `${PIXEL_SCALE * 9}px`,
-            }}
-          />
-        )}
         <img
           src={feederMachineImage}
           className="relative z-0"
